+++ packages/file-renderer/.eslintrc.js
@@ -1,0 +1,4 @@
+module.exports = {
+  root: true,
+  extends: ["@giz/eslint-config"],
+};